Corregido tipo de datos de los id de los modeloos Imagen e ImagenServicio

+++ src/main/java/com/smty/ApiServiciosProfesionales/Models/Imagen.java
@@ -14,7 +14,7 @@ public class Imagen {
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="id_imagen")
-    private int idImagen;
+    private Long idImagen;
 
     @NotEmpty
 	@Column(name="url", length = 100)
